Add config validation call to all armada components (#227)
* Add config validation to all armada components

* Make invalid config error and remove required tag from unrequired fields
