+++ internal/eventingester/configuration/validation.go
@@ -1,0 +1,10 @@
+package configuration
+
+import (
+	"github.com/go-playground/validator/v10"
+)
+
+func (c EventIngesterConfiguration) Validate() error {
+	validate := validator.New()
+	return validate.Struct(c)
+}
