+++ internal/executor/configuration/validation.go
@@ -1,0 +1,8 @@
+package configuration
+
+import "github.com/go-playground/validator/v10"
+
+func (c ExecutorConfiguration) Validate() error {
+	validate := validator.New()
+	return validate.Struct(c)
+}
