+++ internal/scheduler/configuration/validation_test.go
